refactor(calculator): calculator 디렉토리 명 -> model 로 변경
calculator 디렉토리 명을 model로 변경함

변경된 사항:
- calculator -> model 변경
