+++ src/test/java/calculator/model/CalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator.calculator;
+package calculator.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
